+++ src/net/pgfmc/core/inventoryAPI/MenuInventory.java
@@ -1,0 +1,19 @@
+package net.pgfmc.core.inventoryAPI;
+
+public abstract class MenuInventory extends BaseInventory {
+	
+	Button[][] pages;
+	
+	
+	
+
+	public MenuInventory(SizeData size, String name) {
+		super(size, name);
+		
+		
+		
+		
+		
+	}
+
+}
